+++ fiddles/svg/fiddle-0028-BackgroundCarousel/src/metadata.js
@@ -1,3 +1,0 @@
-app.metadata = app.metadata || {
-    gitHubUrl: 'https://github.com/bradyhouse/house/tree/master/fiddles/svg/fiddle-0028-BackgroundCarousel'
-  };