+++ CHANGELOG.markdown
@@ -1154,10 +1154,9 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **SVG**
     1.  Added [fiddles/svg/fiddle-0026-Images](fiddles/svg/fiddle-0026-Images) ~ [Issue 44](https://github.com/bradyhouse/house/issues/44)
     2.  Added [fiddles/svg/fiddle-0027-ImageCloud](fiddles/svg/fiddle-0027-ImageCloud) ~ [Issue 51](https://github.com/bradyhouse/house/issues/51)
-    3.  Started [fiddles/svg/fiddle-0028-BackgroundCarousel](fiddles/svg/fiddle-0028-BackgroundCarousel)
 
 * **Node**
-    1.  Started [fiddles/node/fiddle-0017-WebApp](fiddles/node/fiddle-0017-WebApp) ~ [Issue 46](https://github.com/bradyhouse/house/issues/46)
+    1.  Added [fiddles/node/fiddle-0017-WebApp](fiddles/node/fiddle-0017-WebApp) ~ [Issue 46](https://github.com/bradyhouse/house/issues/46)
     2.  Added [fiddles/node/fiddle-0018-XmlToJson](fiddles/node/fiddle-0018-XmlToJson) ~ [Issue 48](https://github.com/bradyhouse/house/issues/48)
     3.  Added dependency (submodule) to [jimbojw > right-way](https://github.com/jimbojw/right-way)